Refactor function out of main module #24
