+++ pysilcam/__main__.py
@@ -21,7 +21,7 @@ from pysilcam import __version__
 from pysilcam.acquisition import Acquire
 from pysilcam.background import backgrounder
 from pysilcam.config import PySilcamSettings, updatePathLength
-from pysilcam.process import processImage
+from pysilcam.process import processImage, write_stats
 
 if not sys.warnoptions:
     warnings.simplefilter("ignore")
@@ -621,25 +621,6 @@ def collect_rts(settings, rts, stats_all):
         filename = os.path.join(settings.General.datafile,
                                 'OilGasd50.csv')
         rts.to_csv(filename)
-
-
-def write_stats(datafilename, stats_all):
-    '''
-    Writes particle stats into the csv ouput file
-
-    Args:
-        datafilename (str):     filame prefix for -STATS.h5 file that may or may not include a path
-        stats_all (DataFrame):  stats dataframe returned from processImage()
-    '''
-
-    # create or append particle statistics to output file
-    # if the output file does not already exist, create it
-    # otherwise data will be appended
-    # @todo accidentally appending to an existing file could be dangerous
-    # because data will be duplicated (and concentrations would therefore
-    # double) GUI promts user regarding this - directly-run functions are more dangerous.
-    with pd.HDFStore(datafilename + '-STATS.h5', 'a') as fh:
-        stats_all.to_hdf(fh, 'ParticleStats/stats', append=True, mode='r+', format='t', data_columns=True)
 
 
 def check_path(filename):
