+++ pysilcam/process.py
@@ -513,3 +513,22 @@ def processImage(nnmodel, class_labels, image, settings, logger, gui):
         return None
 
     return stats_all
+
+
+def write_stats(datafilename, stats_all):
+    '''
+    Writes particle stats into the csv ouput file
+
+    Args:
+        datafilename (str):     filame prefix for -STATS.h5 file that may or may not include a path
+        stats_all (DataFrame):  stats dataframe returned from processImage()
+    '''
+
+    # create or append particle statistics to output file
+    # if the output file does not already exist, create it
+    # otherwise data will be appended
+    # @todo accidentally appending to an existing file could be dangerous
+    # because data will be duplicated (and concentrations would therefore
+    # double) GUI promts user regarding this - directly-run functions are more dangerous.
+    with pd.HDFStore(datafilename + '-STATS.h5', 'a') as fh:
+        stats_all.to_hdf(fh, 'ParticleStats/stats', append=True, format='t', data_columns=True)
